+++ src/main/java/cz/cvut/fsv/webgama/domain/Coordinate.java
@@ -2,21 +2,21 @@ package cz.cvut.fsv.webgama.domain;
 
 public class Coordinate {
 
-	private Integer id;
-	
+	private Long id;
+
 	private String name;
-	
+
 	private Double x;
-	
+
 	private Double y;
-	
+
 	private Double z;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
